chore(restructure): -re-arrange folder structure [Finishes #165361555]
