+++ package.json
@@ -5,7 +5,7 @@
   "main": "server.js",
   "scripts": {
     "start": "babel-node server/server.js",
-    "test": "nyc --reporter=lcov --reporter=text mocha --require babel-register --timeout 4000 ./tests/*.js --exit",
+    "test": "nyc --reporter=lcov --reporter=text mocha --require babel-register --timeout 4000 ./server/tests/*.js --exit",
     "dev": "node_modules/.bin/nodemon server/server.js --exec babel-node --",
     "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
