+++ server/controllers/auth/signin.controller.js
@@ -4,7 +4,7 @@ import jwt from 'jsonwebtoken';
 import { config } from 'dotenv';
 import User from '../../models/user.model';
 import { setAuthToken } from '../../utils/helpers';
-import validateSignInInput from '../../../validation/authentication/signin';
+import validateSignInInput from '../../validation/authentication/signin';
 
 config();
