+++ server/controllers/auth/signup.controller.js
@@ -2,7 +2,7 @@
 import jwt from 'jsonwebtoken';
 import { config } from 'dotenv';
 import User from '../../models/user.model';
-import validateSignUpInput from '../../../validation/authentication/signup';
+import validateSignUpInput from '../../validation/authentication/signup';
 import { setAuthToken } from '../../utils/helpers';
 
 config();
