+++ server/validation/isEmpty.js
